[1/2] base: Tweaks our stuff by default enable
We got some inconsistencies between our custom tweaks app and the things we really want to have enabled by default.
Because i thought that i could enable stuff on our tweaks app and will be enabled on first boot but it requires that you open it to the stuff getting applied.
As i heard before: Geez noob rom makers, what are they thinking.

This enables the following stuff:
- Icon on toast
- Scroll qs on small qs
- Music title on qs music tile
- Lockscreen Visualizer
- Coverart on LS
- 4G instead of LTE
- Allow LED on DND

Also, this tweaks our following stuff:
- Recents FAB position (to bottom center)

And finally, this disables the following stuff:
- Swap volume button on rotation
- Ticker
- Network traffic
- Headset enables opens music app

Also, thanks to aex team to help me to understand how stuff gets enabled with this commit:
   -https://github.com/AospExtended/platform_packages_apps_Extensions/commit/32341ed45cac3e0dc92ba8dee4253191a84ae462