dumpster: Add AICP Log It
This wonderful tool made by AICP helps people to know how to make the respective logs when something is wrong. Also, with the chance of making it a zip or send it to hastebin.

bootleggers edit: Add a message to explain new people what's the logs, and hope they know what have to send.