Dumpster: The no-no locking v2
This is just a bunch of preventions to users to get their hands on features that won't affect them on various situations. Like in case of not having their lockscreen secured or even thinking on Android Go people who enables it and then whines about features not working. To prevent that, we'll use this big commit locking stuff where it doesn't belong to be tweaked.
All squashed into one commit but it uses some checks for owner info, secured lockscreen, fingerprint and also Go builds because "WHY SHUTDOWN ANIMATIONS DOESNT WORK SUR YOU ROM SUCK"